+++ engine/memory/Pool_Editor.cpp
@@ -80,7 +80,7 @@ void Pool_editorEditBlock(Pool<int> & pool) {
         InputInt("##ReleaseInput", &releaseIndex);
         // limit input
         if (releaseIndex < 0) releaseIndex = 0;
-        if (releaseIndex >= pool._size) releaseIndex = pool._size - 1;
+        if (releaseIndex >= pool._size) releaseIndex = (int)pool._size - 1;
 
         PopItemWidth();
         SameLine();
@@ -110,7 +110,7 @@ void Pool_editorEditBlock(Pool<int> & pool) {
     // array contents
     float colSize = 20.f;
     int nCols = 16;
-    int nRows = pool._size / nCols + (pool._size % nCols != 0);
+    int nRows = (int)pool._size / nCols + (pool._size % nCols != 0);
     int index = 0;
     BeginTable("Pool", nCols);
     for (int row = 0; row < nRows; ++row) {
